connected frontend matching html page with the backend
