+++ django_project/volunteers_r_us/matching/apps.py
@@ -1,0 +1,4 @@
+from django.apps import AppConfig
+class MatchingConfig(AppConfig):
+    name = "volunteers_r_us.matching"
+    label = "matching"
