+++ django_project/volunteers_r_us/matching/data.py
@@ -2,7 +2,7 @@ from .domain import Volunteer, Event
 
 VOLUNTEERS = [
     Volunteer(
-        1,"Alice Nguyen",
+        1,"Allie Nguyen",
         {"Spanish","CPR"},
         {"English","Spanish"},
         {"sat_am","sun_pm"},
